requirements.txt
absl-py==1.4.0
astor==0.8.1
cached-property==1.5.2
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
gast==0.5.4
grpcio==1.48.2
h5py==3.1.0
imageio==2.15.0
importlib-metadata==4.8.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
networkx==2.5.1
nibabel==3.2.2
numpy==1.16.4
packaging==21.3
Pillow==8.4.0
protobuf==3.19.6
pyparsing==3.0.7
python-dateutil==2.8.2
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.16.0
tensorboard==1.12.2
tensorflow-gpu==1.12.2
termcolor==1.1.0
tifffile==2020.9.3
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0